// src/api-client/models/donation-reason.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Why the line was donatable.
 * @export
 * @enum {string}
 */

export const DonationReason = {
    NearExpiry: 'NEAR_EXPIRY',
    DamagedPackaging: 'DAMAGED_PACKAGING',
    SurplusStock: 'SURPLUS_STOCK',
    AestheticImperfection: 'AESTHETIC_IMPERFECTION'
} as const;

export type DonationReason = typeof DonationReason[keyof typeof DonationReason];