// src/api-client/models/product-category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Department as it was.
 * @export
 * @enum {string}
 */

export const ProductCategory = {
    Bakery: 'BAKERY',
    Dairy: 'DAIRY',
    Produce: 'PRODUCE',
    Pantry: 'PANTRY',
    Meat: 'MEAT',
    Beverage: 'BEVERAGE',
    BabyCare: 'BABY_CARE',
    Deli: 'DELI',
    Prepared: 'PREPARED'
} as const;

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];