// src/api-client/models/unit-of-measure.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Unit the quantity is counted in.
 * @export
 * @enum {string}
 */

export const UnitOfMeasure = {
    Unit: 'UNIT',
    Pack: 'PACK',
    Crate: 'CRATE',
    Box: 'BOX',
    Kg: 'KG'
} as const;

export type UnitOfMeasure = typeof UnitOfMeasure[keyof typeof UnitOfMeasure];